EOS-18823: Mini Provisioner: Restructure s3setup_prereqs.json by separating prerequistes sections for each phase 	Change description:             Restructure s3setup_prereqs.json for each phase as described in the ticket's DoD and assocaited code changes in the respective phases and in setupcmd.py         Added/modified/deleted files: 		modified:   scripts/install_prereqs.sh 		modified:   scripts/provisioning/_s3_setup 		modified:   scripts/provisioning/cleanupcmd.py 		modified:   scripts/provisioning/configcmd.py 		modified:   scripts/provisioning/initcmd.py 		modified:   scripts/provisioning/postinstallcmd.py 		modified:   scripts/provisioning/preparecmd.py 		modified:   scripts/provisioning/resetcmd.py 		modified:   scripts/provisioning/s3setup_prereqs.json 		modified:   scripts/provisioning/setupcmd.py 		modified:   scripts/provisioning/testcmd.py         Unit test: 		Full single node mini-provisioner deployment using install_prereqs.sh on local VM with various edits in s3setup_prereqs.json to test corner cases
