+++ scripts/provisioning/_s3_setup
@@ -25,7 +25,6 @@ import argparse
 import traceback
 import errno
 
-from postinstallcmd import PostInstallCmd
 
 def main():
   parser = argparse.ArgumentParser("S3server setup command")
@@ -54,11 +53,12 @@ def main():
 
   args = parser.parse_args()
   try:
-    # run post_install always, the failure will raise exception in PostInstallCmd class
-    PostInstallCmd().process()
-    sys.stdout.write('INFO: post_install validations successful.\n')
-    
-    if args.command == 'cleanup':
+    if args.command == 'post_install':
+      from postinstallcmd import PostInstallCmd
+      PostInstallCmd().process()
+      sys.stdout.write('INFO: post_install successful.\n')
+
+    elif args.command == 'cleanup':
       from cleanupcmd import CleanupCmd
       CleanupCmd(args.config).process()
       sys.stdout.write('INFO: cleanup successful.\n')
