+++ scripts/provisioning/preparecmd.py
@@ -36,3 +36,4 @@ class PrepareCmd(SetupCmd):
   def process(self):
     """Main processing function."""
     sys.stdout.write(f"Processing {self.name} {self.url}\n")
+    self.phase_prereqs_validate(self.name)
