+++ scripts/provisioning/resetcmd.py
@@ -39,6 +39,7 @@ class ResetCmd(SetupCmd):
   def process(self):
     """Main processing function."""
     sys.stdout.write(f"Processing {self.name} {self.url}\n")
+    self.phase_prereqs_validate(self.name)
     try:
       sys.stdout.write('INFO: Cleaning up log files.\n')
       self.CleanupLogs()
